Fixed tail-head implementation to allow tail call optimizations
The fix consist in using the tail iterator returned by an iterator
instead of discarding it and using always the same iterator as was
initially given to us.
This fix will allow tail optimization when implementing concat(): once
all the concat-ed lists except the last one have been consumed we can
return the iterator of the last list instead of the wrapped iterator.
This will also allow efficient unshift/pushBack.
Introcuce a NIL constant that returns an empty list tail.
Multiples values returned in a single iterator call (handled with
_buffer) are disabled for now: we have more basic things to make work
first. An iterator must return a single value.