removing support of perl < 5.16 (temporary?)
Perlude::Open was crappy and i replaced it
by a new implementation of lines() using a new
as_open() which uses &CORE::Open.

then travis reported all the perl < 5.16 to fail.
on the other hand: removing lines() from perlude
core (as it's the wrong place for it) was in the
todolist for a while.

I need to investigate to choose a replacement.